# packages/sdk/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/APIError.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorPayload"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "AdminAuthCreateSessionRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/AdminAuthCreateSessionRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "AdminAuthSessionData": {
        "additionalProperties": false,
        "properties": {
          "adminId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "focusArea": {
            "type": "string"
          },
          "isAuthenticated": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string"
          },
          "lastVisitedRoute": {
            "type": "string"
          }
        },
        "required": [
          "adminId",
          "email",
          "focusArea",
          "isAuthenticated"
        ],
        "type": "object"
      },
      "AdminAuthSessionResponseBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/AdminAuthSessionResponseBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AdminAuthSessionData"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
      "AdminAuthSessionUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/AdminAuthSessionUpdateRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "lastVisitedRoute": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ErrorPayload": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminSessionAuth": {
        "description": "Admin session cookie issued by POST /api/v1/admin/auth/session.",
        "in": "cookie",
        "name": "bidanapp_admin_session",
        "type": "apiKey"
      },
      "CustomerSessionAuth": {
        "description": "Customer session cookie issued by POST /api/v1/customers/auth/session.",
        "in": "cookie",
        "name": "bidanapp_customer_session",
        "type": "apiKey"
      },
      "ProfessionalSessionAuth": {
        "description": "Professional session cookie issued by POST /api/v1/professionals/auth/session.",
        "in": "cookie",
        "name": "bidanapp_professional_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-generated OpenAPI contract for BidanApp.",
    "title": "bidanapp-backend",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/auth/session": {
      "delete": {
        "operationId": "delete-admin-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminSessionAuth": []
          }
        ],
        "summary": "Revoke the authenticated admin session",
        "tags": [
          "Admin Auth"
        ]
      },
      "get": {
        "operationId": "get-admin-auth-session",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthSessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminSessionAuth": []
          }
        ],
        "summary": "Get the authenticated admin session",
        "tags": [
          "Admin Auth"
        ]
      },
      "post": {
        "operationId": "create-admin-auth-session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAuthCreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthSessionResponseBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create an authenticated admin session",
        "tags": [
          "Admin Auth"
        ]
      },
      "put": {
        "operationId": "update-admin-auth-session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminAuthSessionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuthSessionResponseBody"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "AdminSessionAuth": []
          }
        ],
        "summary": "Update authenticated admin session metadata",
        "tags": [
          "Admin Auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}